chores: changing file directory and deleting unused file
